Process WIP labels first
We currently process WIP labels last. The idea being that if labels
changed, we would take those into account, but since we are reading
the label list from the original event data, we don't actually see the
new labels. Do WIP first, and if labels change after, an event will
fire and they will be taken care of and we won't end up overwriting
new data with old data.